Create an abstract MemoryProviderSingleton class. Pass config instead of instantiating a new one where used.
