+++ scripts/memory/base.py
@@ -1,0 +1,34 @@
+import abc
+from config import AbstractSingleton
+import openai
+
+
+def get_ada_embedding(text):
+    text = text.replace("\n", " ")
+    return openai.Embedding.create(input=[text], model="text-embedding-ada-002")["data"][0]["embedding"]
+
+
+def get_text_from_embedding(embedding):
+    return openai.Embedding.retrieve(embedding, model="text-embedding-ada-002")["data"][0]["text"]
+
+
+class MemoryProviderSingleton(AbstractSingleton):
+    @abc.abstractmethod
+    def add(self, data):
+        pass
+
+    @abc.abstractmethod
+    def get(self, data):
+        pass
+
+    @abc.abstractmethod
+    def clear(self):
+        pass
+
+    @abc.abstractmethod
+    def get_relevant(self, data, num_relevant=5):
+        pass
+
+    @abc.abstractmethod
+    def get_stats(self):
+        pass
